:construction: More work on Cython support.
A few features are still needed if we're to generate these files
correctly. While we can dump the output in, guaranteeing it works in all
cases will be difficult. This is simply due to the way in which we build
via cython. What may have to happen is I end up writing a
cython-cc/cython-cxx wrapper that knows how to call out to the various
compilers and can also generate a .d file all in one go. It'd probably
help in the long run, to be quite honest. But that's for another weekend
project.