:sparkles: Add CheckEnvironment module
:sparkles: Added a new ArgParse module. This module is going to save me
a lot of work when it comes to implementing other functions. It's a bit
more complex, but saves me a few extra lines. Still not comfortable with
the naming convention for each of OPTIONS, ONEVALUE, MULTIVALUE args,
but I can probably come up with something better *later*. Right now we
use `OPTIONS`, `VALUES`, and `GROUPS`. This last one might be changed to
`LISTS`.

:truck: Move all CheckXXX modules to be under the CheckEnvironment
module.

This commit, along with the last one, will hopefully show the new
organization going forward. This is simply because it has become too
much work to keep this massive project organized in the way it currently
is. This also gives me the opportunity to begin a larger refactor for
existing modules so that users can still fine tune what exactly they
want or need without having to include TONS of files everytime. This was
also becoming a problem for me, and I needed to solve it sooner, rather
than later.

NOTE: There are still a few more changes needed for CheckEnvironment to
be considered feature complete (for now). Additionally, it might be
renamed to CheckToolchain. Luckily we aren't even at 0.1 yet so we can
still take our time with this.