:boom: :construction: Massive refactor in progress for a 1.0 release
This revision should NOT be pulled unless you're absolutely gungho. This
changes everything and breaks basically everything but the most basic of
workflows.

This should be the final refactor. Several bugs and issues have been
discovered. We're doing more monkey patching to get around CMake's
limitations, and adding support for more tooling so things like
`cmake --build <dir> --target fmt` can be used to invoke clang-format.
Additional tooling will be added over time.

Once this refactor is finished, we'll be in a stable API state, while
documentation will follow. (And then hopefully we'll FINALLY be done)