:boom: Removed more files from the older refactor
:bug: Resolved a few issues.

:pencil2: We might need to actually spend some time figuring out if we
want to namespace *all* our functions, and if we even want to support
monkey patching. If we monkey patch, our files can stay mostly the same.
However, there are some edge cases where things break down and "doing
the right thing" isn't worth it, so we end up not monkey patching at
all.

This is problematic, but then again, so is having an ixm_ in front of
every single function a user needs to write. It becomes more obvious of
a tooling split, when ideally people should be able to keep writing what
they're writing with little interference from us.

There is of course the current situation of the DefaultLayout
1) Not actually generating tests
2) Not automatically adding dependencies to the correct target.

I don't know if this will be possible. I'd like to think it is, but at the end
of the day, it's CMake and it's all broken to hell anyhow. It just makes
me sad that I can't go in and just ADD these things in the native code.
It's *extremely* frustrating. We're also going to be keeping track of SO
MUCH stuff. I'm not very happy about it, and it makes the chance of
something going wrong VERY LIKELY.