:recycle: Massive refactor regarding APIs
:tada: getting closer to an actual alpha release! That's when the initial API
set will have been declared stable, and will no longer be iterated on or
changed. Only newer features or capabilities will be added.

:boom: genex() has been renamed to genexp()
:boom: add_executable() override has been removed so we can work with
vcpkg. I don't want to step on their toes.

:sparkles: Initial "event" variables have been added for when targets
are added. The actual event system must still be designed for ease of
use.

Redesigning how Blueprints are discovered, as well as where they are
located within the project. This is still a work in progress, however
existing behavior should mostly remain.

Also, I'm attempting to clean up and reorganize some operations, such
that there will be a better API for generating files.

Events are also going to