:fire: Removed CMake Custom Language support for Rust
Unfortunately, Rust has a mildly different compilation model and this
requires some work to be done at the CMake level to support compiling a
Rust crate and linking it to a C or C++ target. This also means we
probably won't be able to have golang support, but it remains to be
seen.

:sparkles: Add initial Android Toolchain support.

This falls in line with only requiring a user to define
`CMAKE_SYSTEM_NAME` and (optionally) `CMAKE_SYSTEM_PROCESSOR` to cross
compile for Android.

:sparkles: Add support for FindBullet.cmake in a *modern* way

The current CMake FindBullet needs serious work, while the IXM version
is just 4 lines of text :)

:sparkles: Add FindRust.cmake

We'll be using this to allow IXM to "enable" Rust support in a build.

:bug: Removed commented FindBullet code from SCCache

It's been there forever and needed to be removed

:bug: The QNX Toolchain will now FATAL_ERROR if the QNX_HOME environment
variable hasn't been set. This indicates that the user hasn't invoked
the QNX environment script.