:fire: Removed custom executable/library functions
They will be replaced with a more generic and ubiquitous `target()`
function in the near future before the alpha release.

:recycle: Changed how toolchains are detected.

Currently a foreach is being used, but we'll be moving to a "property based"
solution in the near future.

:recycle: Began renaming all properties to use the
ixm::<namespace/feature>::<property> syntax. Some namespacing has to be
decided on, so some things will be broken for a small amount of time.

:truck: Move all CMake override commands to API::CMake.

This allows us to know at a glance what builtin CMake commands are being
overwritten.

:sparkles: Add support for defining subcommands via properties

This will allow functions to be redirected without having to be
redefined. Documentation regarding this approach will have to be written
but it's intended that this is how users can hook custom functions into
a given operation.

:sparkles: Move all initial property settings to API/Prelude. We may
also place all "non-project" core functions into this Submodule (i.e.,
underneath a `Prelude::*` import.

:sparkles: `locals()` command learned to filter variables in the written
to variable.

:sparkles: Initial settings for "builtin" events (e.g.,
CMAKE_FIND_PACKAGE_NAME and CMAKE_CURRENT_LIST_DIR) have been
implemented. We do not have a set in stone way to register for events
yet, but that will be implemented soon.

:sparkles: `inspect()` macro is now a function. This has not had an
adverse effect on existing code and also gives better performance.

:sparkles: Add initial FindMono.cmake implementation

:sparkles: Add initial FindSDL2.cmake implementation

:loud_sound: Replacing manual print calls with log(<SUBCOMMAND>) calls.
While logging itself is not yet *entirely* finished, it will be soon.