:sparkles: Introduce new `attribute()` command.
This will allow us to have a better lookup strategy. We're effectively
hacking a hierarchy into the GLOBAL PROPERTY system (instead of
directory properties). This way properties are ONLY for projects, and
IXM. (Blueprints might be added to the lookup later)
This command will also be used as the primary interaction with
customization points for users. It is, effectively, the API mentioned in #20

attribute() is a lot more "fun" to use, I would say. It's API is
consistent with IXM, though its naming conventions still need some
wwork.

:pencil: Left notes on what could be done to support other targets, such
as Cython, Node, etc.

:bug: Fix dict(JSON) serialization bug when trying to "check" if a
string is actually a number.
:bug: Fix `inspect()` continuously appending CMAKE_MESSAGE_INDENT if
more than one parameter was passed to it.

:poop: parse will now set max to 9 if there is not ixm::parse::max
property. This will be adjusted later.

:truck: Rename var() to assign() (Closes #23)
  This function is used like
    `assign(<var> ? LOOKUP1 LOOKUP2 : "DEFAULT" "VALUES")`
  It is MUCH more powerful than `var()` ever was, and also has *mildly*
  similar syntax to a C style ternary operation. Extremely mild. We
  don't do a condition check, effectively.

:truck: Moved all functions used to implement libraries into API::Author.
  This might be changed later to API::Core, but we'll see. It would
  require the GitHub issue labels to be updated as we call IXM::API the
  "core modules".

:truck: Duplicated all properties to use attribute. (#25)
  Quite a bit of work for the customization points is still needed so
  we'll be in a "dual" mode for a while until this is resolved.

:truck: Move CPackConfig generator to `API::Generate`
:truck: Move protobuf stub to `API::Generate`

:construction: Migrating get_property to attribute()
:construction: target_sources override should be moved to target(SOURCES)
:construction: Stub out `define_property` override (#22).
  This may be removed in favor of `attribute(DEFINE)`
:construction: Stub out `target()` (#18)
:construction: Moving old output messages to log()
:construction: Introspection API (#19) is now available by default. This is
simply the `inspect` and `locals()` command. Several other commands will
be added in the future.

:fire: Remove add_executable override
  (use target(PROGRAM) once available)

:fire: Remove plenty of empty or unused files