:sparkles: Add dict(ASSIGN), dict(APPEND), dict(CONCAT)
These replace the (now removed) `dict(INSERT)`. This is better for
performance (see #29), as we aren't spending too much time figuring out
what action to do.

:sparkles: Add dict(FILTER) and dict(TRANSFORM) support

These act like the `list(FILTER)` and `list(TRANSFORM) commands, except
they act on the values of a key. There are probably some strange edge
cases hanging around, but we'll find them in due time.

:alembic: Willing to mark the `dict()` "binary" file format as 'stable'.

It might be changed in the future to permit dumping multiple
dictionaries into a file, but we'll increment the version if that ever
happens.

More testing is needed of course, but that will come in due time.

:art: Clean up some of the variable naming

I understand it might behoove me to be paranoid about CMake's (IMO
inconsistent scoping rules), but I think we can relax most of the issues
I was trying to solve. Thus, I've removed a bunch of the `@` characters
littered throughout the codebase.

:recycle: Refactor internals of `dict()` command to not be as recursive
:recycle: Update internals of IXM to use the newer `dict()` commands

:bug: Fix `log(FATAL)` failing to print.

The performance of the log() function took a hit, as this fix is just a
bit of a hack. We'll need to come up with some way to cache the colors
in a safe manner. Additionally, we need a better way to work with the
CMake 3.15 message() changes.

:fire: `dict(MERGE)` and `dict(SET) have been removed.

These were added for `fetch()` to make passing a `dict()` as a "kwargs"
argument easier. They weren't used internally, and the performance was
atrocious when used. Since `fetch()` is due for an API upgrade prior to
release, I figured it best to remove it now.