:recycle: Quality of Life fixes
:zap: Check() template will now not waste time checking if a variable
was defined. This was a waste of if-else statements.

:fire: genexp() has been removed. We can use string(CONCAT) 9/10 times
when set() doesn't do the job. I am kicking myself for wasting so much
time getting this sort of thing to work.

:fire: glob() has been removed. This lets us close #15

:sparkles: Add mild compatibility layer with CMake 3.16's UNITY_BUILD feature.
We do a few things differently for ourselves because we need more
granular control, *but*, moving to using our UNITY settings vs CMake's
won't be too much of a hop :D

:sparkles: Add preliminary support to Coven for disabling directories
being compiled at all. This is absolutely necessary to have "platform
specific" projects. In our case, we set the `WHEN` on a given target.
The `WHEN` value must be a generator expression that resolves to a
boolean. i.e., `$<BOOL:${VAR}>` or `$<PLATFORM_ID:${name}>`. An example
would be blocking a directory from compiling on Windows

```cmake
set_property(
  TARGET ${PROJECT_NAME}
  PROPERTY
    WHEN $<NOT:$<PLATFORM_ID:Windows>>)
```

:sparkles: Target programs inside of `bin` are now exported with the
same name as their output_name by default.

:sparkles: Backport 3.16's `target_precompile_headers`

:sparkles: Add `PATCH`, `POLICIES`, `OPTIONS` as manual options to
fetch()

:art: Cleaned up the aspect() command's internals. We didnt need all
that "protection"

:bug: Added a few guards regarding color output on windows.

:bug: Fix some logging output not erroring correctly

:bug: Fix logging to files not having a file to output to in the event
that IXM has logged an issue *before* the first call to `project()`

:fire: Delete IXM::Common::* modules

:poop: We're going to be supporting CTest "proper" for a while. I'd love
to write a better test runner, but the way things are going, we're going
to have to delay the CTest/enable_testing() override that will let us
generate a proper CTestConfig.cmake for a project at generation time.
This will also let us make sure that the build is run *before* any tests
are run.