:alembic: Modify how Coven blueprint's targets are implemented
We now *always* generate an INTERFACE target with the PROJECT_NAME. This
is the "public" interface for the project, and is intended to specify
the transitive linking for other projects. Any target(<action>) or
target_<command> from CMake is intended to be used on this target.

We also generate a PROJECT_NAME::private INTERFACE IMPORTED target. This
is for private settings, which works in practice because INTERFACE
IMPORTED targets aren't actually exported (and when 'importing' a
project via find_package, non INTERFACE settings wouldn't really be
present anyhow. I'm fine with this.

Currently we're disabling the 'src/<directory>/main' -> generated binary
rule. I am copying it from Rust's Cargo, but I really do despise it as a
concept. Such is life

Next, the 'main' program is now labeled PROJECT_NAME-program, and is
also aliased to PROJECT_NAME::program. This is also what the 'target' is
currently exported as. If users want to provide a different name,
they're more than free to overwrite the EXPORT_NAME, and create an
additional alias.

Additionally, the generated code library is named PROJECT_NAME-library,
and is also aliased to PROJECT_NAME::library. If it exists, it is *also*
ALIAS'd to PROJECT_NAME::PROJECT_NAME. Otherwise, the PROJECT_NAME
interface library is set to this. OUTPUT_NAME is set correctly on all
targets to prevent unnecessary naming conventions and issues.

We still do not correctly export all targets in a project, nor do we
actually generate a proper <package>-config.cmake file (One that takes
components and set_and_check into account, though we really shouldn't be
using that anyhow. It's a bit too overly unnecessary for the average
project)

:bug msvc-runtime backport has been fixed to support clang-cl, and to be
shorter and more readable.