♻ Refactor Blueprints/Coven/Project to be more correct
There's still a few things left to do.

For starters, we need to rename this file to Targets.cmake. This only
creates the targets, sets up the initial settings, and then finishes.
This isn't project wide for the initialization, and some additional
things like installing, setting components, etc. are needed.

Additionally, the ability to detect *and* build directories that hold a
main.<ext> file is still not implemented. This *is* needed for
generating additional binaries that are to be installed or exported for
the program.

Lastly, a solution is needed for how tools (subprojects/binaries that
are ALWAYS built by the project in a "native" way, even when cross
compiling) are going to interact with the Coven layout. We don't
currently have a single path for "here's my 'build scripts', run this
separate project via try_compile, then import all its exported objects
as if they had been installed". It's frustrating. You'd think the CMake
devs would have a better solution than "lmao, just use an emulator
dummy", but then again this is the same set of developers that think
filesystem entries can be represented as a list of doubles in JSON, and
don't understand how to actually export JSON in a usable format. Nor do
they know how to write a fucking programming language but that ship has
sailed.

I shouldn't be a leading expert in CMake. I'm self taught, mentally ill, and
without a degree and somehow I still run circles around most of their devs.
What the fuck is their excuse? It's so disheartening. 😟