✨ Add source_file_extensions to Introspection API
This simply returns a list of valid (at the time of the call) file
extensions that can be added via `target_sources`.

🔥 Delete most of target() command

It's been mostly superfluous and also was leading to some unnecessary
overhead. After giving it some thought, I'd prefer to not automatically
try to generate source files from things like protobuf automatically.
This would make upgrading users to coven more difficult, and it reduces
IXM's flexibility for the stable alpha. Instead, going forward, we'll
have a target_<name>_sources command placed in the packages we support
(protobuf if possible, cython, etc). Most of the work IXM will be doing
in the future is mostly property based and this just makes it easier to
be flexible (and get out of users ways)

⚡ Working on some optimizations for parsing arguments.

As nice as parse() is, in some cases it isn't needed and in others the
double cmake_parse_arguments call is a bit too heavy. Let's leave it for
users, but try to reduce out dependency on it.

TODO: Still need to tackle the Modules/Use directory. Sphinx settings
need to go into the `FindSphinx` API. Generating documentation is still
an issue for Coven via IXM and it's going to be a while before we can
really tackle the problem.